src/requirements.txt
django==5.1
djangorestframework==3.15.2
python-dotenv==1.0.0
# psycopg2-binary>=2.9.6
uvicorn==0.23.1
gunicorn==22.0.0
django-filter==23.2
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
